feat: set news to /news with Await-based pagination
