+++ website-frontend/src/routes/news/+page.ts
@@ -1,0 +1,8 @@
+/** @type {import('./$types').PageLoad} */
+export async function load({ data }) {
+	return {
+		news_limit: data.news_limit,
+		news_count: data.news_count,
+		news: data.news
+	};
+}
